refactor(middleware): organize middleware, remove MergeRoutes, CtxAborted and remove jape import
