+++ middleware/middleware.go
@@ -1,254 +1,10 @@
 package middleware
 
 import (
-	"context"
-	"errors"
 	"github.com/LumeWeb/portal/core"
-	"net/http"
-	"slices"
-	"strconv"
-	"strings"
-
-	"github.com/casbin/casbin/v2"
 	"github.com/golang-jwt/jwt/v5"
-	"go.sia.tech/jape"
+	"slices"
 )
-
-type AuthTokenContextKeyType string
-type UserIdContextKeyType string
-
-const DEFAULT_USER_ID_CONTEXT_KEY UserIdContextKeyType = "user_id"
-const AUTH_TOKEN_CONTEXT_KEY AuthTokenContextKeyType = "auth_token"
-
-type FindAuthTokenFunc func(r *http.Request) string
-
-func FindAuthToken(r *http.Request, cookieName string, queryParam string) string {
-	authHeader := ParseAuthTokenHeader(r.Header)
-
-	if authHeader != "" {
-		return authHeader
-	}
-
-	if cookie, err := r.Cookie(cookieName); cookie != nil && err == nil {
-		return cookie.Value
-	}
-
-	if cookie, err := r.Cookie(core.AUTH_COOKIE_NAME); cookie != nil && err == nil {
-		return cookie.Value
-	}
-
-	return r.FormValue(queryParam)
-}
-
-func ParseAuthTokenHeader(headers http.Header) string {
-	authHeader := headers.Get("Authorization")
-	if authHeader == "" {
-		return ""
-	}
-
-	authHeader = strings.TrimPrefix(authHeader, "Bearer ")
-	authHeader = strings.TrimPrefix(authHeader, "bearer ")
-
-	return authHeader
-}
-
-type AuthMiddlewareOptions struct {
-	Context        core.Context
-	FindToken      FindAuthTokenFunc
-	Purpose        core.JWTPurpose
-	AuthContextKey string
-	EmptyAllowed   bool
-	ExpiredAllowed bool
-}
-
-func AuthMiddleware(options AuthMiddlewareOptions) func(http.Handler) http.Handler {
-	config := options.Context.Config()
-
-	if options.AuthContextKey == "" {
-		options.AuthContextKey = string(DEFAULT_USER_ID_CONTEXT_KEY)
-	}
-
-	domain := config.Config().Core.Domain
-
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			authToken := options.FindToken(r)
-
-			if authToken == "" {
-				if !options.EmptyAllowed {
-					http.Error(w, "Invalid JWT", http.StatusUnauthorized)
-					return
-				}
-				next.ServeHTTP(w, r)
-				return
-			}
-
-			var audList *jwt.ClaimStrings
-
-			claim, err := core.JWTVerifyToken(authToken, domain, options.Context.Config().Config().Core.Identity.PrivateKey(), func(claim *jwt.RegisteredClaims) error {
-				aud, _ := claim.GetAudience()
-
-				audList = &aud
-
-				if options.Purpose != core.JWTPurposeNone && !jwtPurposeEqual(aud, options.Purpose) {
-					return core.ErrJWTInvalid
-				}
-
-				return nil
-			})
-
-			if err != nil {
-				unauthorized := true
-				if errors.Is(err, jwt.ErrTokenExpired) && options.ExpiredAllowed {
-					unauthorized = false
-				}
-
-				if !unauthorized && audList == nil {
-					if audList == nil {
-						var claim jwt.RegisteredClaims
-
-						unverified, _, err := jwt.NewParser().ParseUnverified(authToken, &claim)
-						if err != nil {
-							http.Error(w, err.Error(), http.StatusInternalServerError)
-							return
-						}
-
-						audList, err := unverified.Claims.GetAudience()
-						if err != nil {
-							http.Error(w, err.Error(), http.StatusInternalServerError)
-							return
-						}
-
-						if jwtPurposeEqual(audList, options.Purpose) {
-							unauthorized = true
-						}
-
-					}
-				}
-
-				if unauthorized {
-					http.Error(w, err.Error(), http.StatusUnauthorized)
-					return
-				}
-			}
-
-			if claim == nil && options.ExpiredAllowed {
-				next.ServeHTTP(w, r)
-				return
-			}
-
-			userId, err := strconv.ParseUint(claim.Subject, 10, 64)
-
-			if err != nil {
-				http.Error(w, core.ErrJWTInvalid.Error(), http.StatusBadRequest)
-				return
-			}
-
-			exists, _, err := options.Context.Services().User().AccountExists(uint(userId))
-
-			if !exists || err != nil {
-				http.Error(w, core.ErrJWTInvalid.Error(), http.StatusBadRequest)
-				return
-			}
-
-			ctx := context.WithValue(r.Context(), UserIdContextKeyType(options.AuthContextKey), uint(userId))
-			ctx = context.WithValue(ctx, AUTH_TOKEN_CONTEXT_KEY, authToken)
-			r = r.WithContext(ctx)
-
-			next.ServeHTTP(w, r)
-		})
-	}
-}
-
-type AuthzOptions struct {
-	Context core.Context
-	Casbin  *casbin.Enforcer
-	Role    string
-}
-
-func AuthzMiddleware(options AuthzOptions) func(http.Handler) http.Handler {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			user := r.Context().Value(DEFAULT_USER_ID_CONTEXT_KEY)
-
-			deny := func() {
-				http.Error(w, "Unauthorized", http.StatusUnauthorized)
-			}
-
-			exists, m, err := options.Context.Services().User().AccountExists(user.(uint))
-			if err != nil || !exists {
-				deny()
-				return
-			}
-
-			if options.Role != m.Role {
-				deny()
-				return
-			}
-
-			ok, err := options.Casbin.Enforce(m.Role, r.URL.Path, r.Method)
-			if err != nil || !ok {
-				deny()
-				return
-			}
-
-			next.ServeHTTP(w, r)
-		})
-	}
-}
-
-func MergeRoutes(routes ...map[string]jape.Handler) map[string]jape.Handler {
-	merged := make(map[string]jape.Handler)
-
-	for _, route := range routes {
-		for k, v := range route {
-			merged[k] = v
-		}
-	}
-
-	return merged
-}
-
-func GetUserFromContext(ctx context.Context, key ...string) uint {
-	realKey := ""
-
-	if len(key) > 0 {
-		realKey = key[0]
-	}
-
-	if realKey == "" {
-		realKey = string(DEFAULT_USER_ID_CONTEXT_KEY)
-	}
-
-	realKeyCtx := UserIdContextKeyType(realKey)
-
-	userId, ok := ctx.Value(realKeyCtx).(uint)
-
-	if !ok {
-		panic("user id stored in context is not of type uint")
-	}
-
-	return userId
-}
-
-func GetAuthTokenFromContext(ctx context.Context) string {
-	authToken, ok := ctx.Value(AUTH_TOKEN_CONTEXT_KEY).(string)
-
-	if !ok {
-		panic("auth token stored in context is not of type string")
-	}
-
-	return authToken
-}
-
-func CtxAborted(ctx context.Context) bool {
-	select {
-	case <-ctx.Done():
-		return true
-	default:
-		return false
-	}
-}
 
 func jwtPurposeEqual(aud jwt.ClaimStrings, purpose core.JWTPurpose) bool {
 	return slices.Contains[jwt.ClaimStrings, string](aud, string(purpose))
